feat(one-line-commands): moves the get_job_wrapper() manager to own file
Clarifying that we have job-function variants move the selection
function to its own file.
