+++ runem/job_execute.py
@@ -7,7 +7,7 @@ from datetime import timedelta
 from timeit import default_timer as timer
 
 from runem.config_metadata import ConfigMetadata
-from runem.job_wrapper_python import get_job_wrapper
+from runem.job_wrapper import get_job_wrapper
 from runem.log import log
 from runem.types import (
     FilePathList,
